requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.4
decorator==4.0.11
defusedxml==0.4.1
Django==1.9.1
django-activity-stream==0.6.3
django-allauth==0.31.0
django-cors-headers==2.0.2
django-encrypted-fields==1.1.2
django-extensions==1.3.11
django-fieldsignals==0.2.3
django-jsonfield==1.0.1
django-rest-auth==0.9.1
django-rest-encrypted-lookup==0.10.1
django-taggit==0.22.1
django-taggit-serializer==0.1.5
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
idna==2.5
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.0
psycopg2==2.7.1
pyasn1==0.2.3
pycrypto==2.6.1
PyJWT==1.4.2
pyparsing==2.2.0
python-keyczar==0.71rc0
python-openid==2.2.5
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
requirements==0.1
six==1.10.0
urllib3==1.21.1
validators==0.11.3
gunicorn
